refactor(util): rename utils folder to util
