+++ src/providers/fal.ts
@@ -4,7 +4,7 @@ import { getEnvString } from '../envars';
 import logger from '../logger';
 import type { ApiProvider, CallApiContextParams, ProviderResponse } from '../types';
 import type { EnvOverrides } from '../types/env';
-import { ellipsize } from '../utils/text';
+import { ellipsize } from '../util/text';
 
 type FalProviderOptions = {
   apiKey?: string;
