+++ src/table.ts
@@ -2,7 +2,7 @@ import chalk from 'chalk';
 import Table from 'cli-table3';
 import { TERMINAL_MAX_WIDTH } from './constants';
 import { ResultFailureReason, type EvaluateTable } from './types';
-import { ellipsize } from './utils/text';
+import { ellipsize } from './util/text';
 
 export function generateTable(
   evaluateTable: EvaluateTable,
